+++ config/routes.rb
@@ -5,6 +5,7 @@ Que::View::Engine.routes.draw do
     post :reschedule_all, on: :collection
     delete :destroy_all, on: :collection
   end
+  resources :queue_metrics, only: %i[index]
 
   root 'welcome#index'
 end