+++ lib/que/view.rb
@@ -34,7 +34,7 @@ module Que
     end
 
     # Public: All the methods delegated to instance. These should match the interface of Que::View::DSL.
-    def_delegators :instance, :fetch_dashboard_stats, :fetch_que_lockers,
+    def_delegators :instance, :fetch_dashboard_stats, :fetch_que_lockers, :fetch_queue_metrics,
                    :fetch_queue_names, :fetch_job_names,
                    :fetch_running_jobs, :fetch_failing_jobs, :fetch_scheduled_jobs, :fetch_finished_jobs,
                    :fetch_expired_jobs, :fetch_job,