+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - rendering running jobs
+- pagination for jobs list page
 
 ## [0.2.3] - 2023-12-01
 ### Fixed