+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.5] - 2024-04-13
+### Fixed
+- bug with rendering jobs without arguments
+
 ## [0.3.4] - 2024-03-07
 ### Modified
 - rendering failing jobs list when they are expired